requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
tekore==4.3.0
yarl==1.7.2
